+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Converter for Croatia (`hr`)
 - Converter for Germany, Mecklenburg-Western Pomerania
+- Converter for Germany, Saarland
 - Converter for Germany, Saxony
 
 ### Changed